arm64/dts: Early pKVM module loading of exynos-pd,pkvm_s2mpu
Previously, kvm-arm.protected_modules was set in the cmdline without any
argument. This corresponds to enabling late module loading for pKVM,
i.e. loaded by userspace and the registration closed by the module
itself.

Recently, early loading was introduced for pKVM modules. It is enabled
by the same cmdline option but with a list of modules as an argument. In
that case, loading is done by the kernel itself before host deprivilege,
thereby reducing the attack surface.

Switch pkvm_s2mpu to early loading in preparation for the deprecation of
the old method. Since pkvm_s2mpu relies on exynos-pd, we also have have
to load exynos-pd using the same method.

Bug: 254835242
Change-Id: I25e69dcdc85a04be0468afe604452ec106da583c
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>
Signed-off-by: Daniel Mentz <danielmentz@google.com>